requirements.txt
nltk
numpy
fbchat
requests
python-dotenv
